requirements.txt
pygame
PyOpenGL
imgui
trimesh